Cast to int

+++ db/block_db.h
@@ -43,7 +43,7 @@ public:
     lseek(fd, 0, SEEK_SET);
     while (remaining > 0)
     {
-        auto remaining_size = std::min(MAX_POSIX_MEMALIGN_SIZE, static_cast<int>(storage_size));
+        auto remaining_size = std::min(MAX_POSIX_MEMALIGN_SIZE, static_cast<int64_t>(storage_size));
         assert(write(fd, buf, remaining_size) != -1);
         remaining -= MAX_POSIX_MEMALIGN_SIZE;
     }
